# openpecha_api/api-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenPecha API v2",
    "version": "2.11.0"
  },
  "paths": {
...
    "/v2/editions/{edition_id}/recordings": {
      "get": {
        "tags": [
          "Editions"
        ],
        "summary": "Get edition recordings",
        "description": "Retrieve all audio recordings of an edition.",
        "operationId": "get_recordings_v2_editions__edition_id__recordings_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "edition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the edition",
              "title": "Edition Id"
            },
            "description": "The ID of the edition"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecordingOutput"
                  },
                  "title": "Response Get Recordings V2 Editions  Edition Id  Recordings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Editions"
        ],
        "summary": "Add edition recording",
        "description": "Upload an audio recording of an edition, with its metadata as a JSON form field.",
        "operationId": "post_recording_v2_editions__edition_id__recordings_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "edition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the edition",
              "title": "Edition Id"
            },
            "description": "The ID of the edition"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_recording_v2_editions__edition_id__recordings_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/persons": {
      "get": {
        "tags": [
          "Persons"
        ],
        "summary": "List all persons",
        "description": "Retrieve a paginated list of persons.",
        "operationId": "get_all_persons_v2_persons_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/models__base__NonEmptyStr__2"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by person name, minimum 2 characters",
              "title": "Name"
            },
            "description": "Filter by person name, minimum 2 characters"
          },
          {
            "name": "bdrc",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by BDRC ID",
              "title": "Bdrc"
            },
            "description": "Filter by BDRC ID"
          },
          {
            "name": "wiki",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Wiki ID",
              "title": "Wiki"
            },
            "description": "Filter by Wiki ID"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of items to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_PersonOutput_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Persons"
        ],
        "summary": "Create a person",
        "description": "Create a new person.",
        "operationId": "create_person_v2_persons_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/recordings/{recording_id}": {
      "get": {
        "tags": [
          "Recordings"
        ],
        "summary": "Get a recording",
        "description": "Retrieve metadata for an audio recording.",
        "operationId": "get_recording_v2_recordings__recording_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the recording",
              "title": "Recording Id"
            },
            "description": "The ID of the recording"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Recordings"
        ],
        "summary": "Update a recording",
        "description": "Partially update a recording's metadata.",
        "operationId": "update_recording_v2_recordings__recording_id__patch",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the recording",
              "title": "Recording Id"
            },
            "description": "The ID of the recording"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordingPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recordings"
        ],
        "summary": "Delete recording",
        "description": "Delete a recording's metadata and its stored audio file.",
        "operationId": "delete_recording_v2_recordings__recording_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the recording",
              "title": "Recording Id"
            },
            "description": "The ID of the recording"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIContribution-Input": {
        "properties": {
          "type": {
            "type": "string",
            "const": "ai",
            "title": "Type"
          },
          "id": {
            "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
          },
          "role": {
            "$ref": "#/components/schemas/ContributorRole"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "id",
          "role"
        ],
        "title": "AIContribution"
      },
      "AIContribution-Output": {
        "properties": {
          "type": {
            "type": "string",
            "const": "ai",
            "title": "Type"
          },
          "id": {
            "$ref": "#/components/schemas/NonEmptyStr-Output"
          },
          "role": {
            "$ref": "#/components/schemas/ContributorRole"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "id",
          "role"
        ],
        "title": "AIContribution"
      },
...
      "AudioFormat": {
        "type": "string",
        "enum": [
          "mp3",
          "wav",
          "m4a",
          "ogg",
          "flac"
        ],
        "title": "AudioFormat"
      },
...
      "Body_post_recording_v2_editions__edition_id__recordings_post": {
        "properties": {
          "metadata": {
            "type": "string",
            "title": "Metadata",
            "description": "JSON-encoded recording metadata"
          },
          "audio": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio",
            "description": "Audio file"
          }
        },
        "type": "object",
        "required": [
          "metadata",
          "audio"
        ],
        "title": "Body_post_recording_v2_editions__edition_id__recordings_post"
      },
...
      "ContributionInputItem": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PersonContributionInput"
          },
          {
            "$ref": "#/components/schemas/AIContribution-Input"
          }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "ai": "#/components/schemas/AIContribution-Input",
            "person": "#/components/schemas/PersonContributionInput"
          }
        }
      },
      "ContributionOutputItem": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PersonContributionOutput"
          },
          {
            "$ref": "#/components/schemas/AIContribution-Output"
          }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "ai": "#/components/schemas/AIContribution-Output",
            "person": "#/components/schemas/PersonContributionOutput"
          }
        }
      },
      "ContributorRole": {
        "type": "string",
        "enum": [
          "translator",
          "reviser",
          "author",
          "scholar",
          "narrator"
        ],
        "title": "ContributorRole"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id"
        ],
        "title": "IdResponse"
      },
...
      "LicenseType": {
        "type": "string",
        "enum": [
          "cc0",
          "public",
          "cc-by",
          "cc-by-sa",
          "cc-by-nd",
          "cc-by-nc",
          "cc-by-nc-sa",
          "cc-by-nc-nd",
          "copyrighted",
          "unknown"
        ],
        "title": "LicenseType"
      },
      "LocalizedString-Input": {
        "additionalProperties": {
          "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
        },
        "type": "object",
        "minProperties": 1,
        "title": "LocalizedString"
      },
      "LocalizedString-Output": {
        "additionalProperties": {
          "$ref": "#/components/schemas/NonEmptyStr-Output"
        },
        "type": "object",
        "minProperties": 1,
        "title": "LocalizedString"
      },
      "NonEmptyStr-Output": {
        "type": "string",
        "minLength": 1,
        "pattern": "\\S"
      },
...
      "PaginatedResponse_PersonOutput_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PersonOutput"
            },
            "type": "array",
            "title": "Items"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "has_more",
          "offset",
          "limit"
        ],
        "title": "PaginatedResponse[PersonOutput]"
      },
...
      "PersonContributionInput": {
        "properties": {
          "type": {
            "type": "string",
            "const": "person",
            "title": "Type"
          },
          "id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "bdrc_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/ContributorRole"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "role"
        ],
        "title": "PersonContributionInput"
      },
      "PersonContributionOutput": {
        "properties": {
          "type": {
            "type": "string",
            "const": "person",
            "title": "Type"
          },
          "id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "bdrc_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/ContributorRole"
          },
          "name": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalizedString-Output"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "role"
        ],
        "title": "PersonContributionOutput"
      },
      "PersonInput": {
        "properties": {
          "bdrc": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "wiki": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "$ref": "#/components/schemas/LocalizedString-Input"
          },
          "alt_names": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LocalizedString-Input"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Names"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PersonInput"
      },
      "PersonOutput": {
        "properties": {
          "bdrc": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "wiki": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "$ref": "#/components/schemas/LocalizedString-Output"
          },
          "alt_names": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LocalizedString-Output"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Names"
          },
          "id": {
            "$ref": "#/components/schemas/NonEmptyStr-Output"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "PersonOutput"
      },
...
      "RecordingOutput": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalizedString-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "license": {
            "$ref": "#/components/schemas/LicenseType",
            "default": "public"
          },
          "date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NonEmptyStr-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "id": {
            "$ref": "#/components/schemas/NonEmptyStr-Output"
          },
          "edition_id": {
            "$ref": "#/components/schemas/NonEmptyStr-Output"
          },
          "text_id": {
            "$ref": "#/components/schemas/NonEmptyStr-Output"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/ContributionOutputItem"
            },
            "type": "array",
            "title": "Contributions"
          },
          "format": {
            "$ref": "#/components/schemas/AudioFormat"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "edition_id",
          "text_id",
          "contributions",
          "format",
          "size_bytes"
        ],
        "title": "RecordingOutput"
      },
      "RecordingPatch": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalizedString-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "license": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LicenseType"
              },
              {
                "type": "null"
              }
            ]
          },
          "date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/models__base__NonEmptyStr__1"
              },
              {
                "type": "null"
              }
            ]
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "contributions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ContributionInputItem"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Contributions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RecordingPatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "models__base__NonEmptyStr__1": {
        "type": "string",
        "minLength": 1,
        "pattern": "\\S"
      },
      "models__base__NonEmptyStr__2": {
        "type": "string",
        "minLength": 2,
        "pattern": "\\S"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}